backend/requirements.txt
fastapi
uvicorn
supabase
pandas
openpyxl
shapely
scipy
numpy
pydantic
python-dotenv
geojson
python-multipart
